+++ src/main/java/com/projet/scootop/scootop/controller/ApiProviderController.java
@@ -1,12 +1,7 @@
 package com.projet.scootop.scootop.controller;
 
-import com.projet.scootop.scootop.domain.provider.Cameraman;
-import com.projet.scootop.scootop.domain.provider.WearableSociety;
-import com.projet.scootop.scootop.model.ClubDTO;
 import com.projet.scootop.scootop.model.provider.*;
 import com.projet.scootop.scootop.service.provider.*;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -36,7 +31,7 @@ public class ApiProviderController {
         return cameramanService.getAll();
     }
     @GetMapping("api/cameraman/{id}")
-    CameramanDTO getCameraman(@PathVariable int id){
+    CameramanDTO getCameraman(@PathVariable Long id){
         return cameramanService.get(id);
     }
     @PostMapping("api/cameraman")
@@ -45,12 +40,12 @@ public class ApiProviderController {
         return cameraman;
     }
     @PutMapping("api/cameraman/{id}")
-    String updateCameraman(@RequestBody String body,@PathVariable int id) {
+    String updateCameraman(@RequestBody String body,@PathVariable Long id) {
         String cameraman = body;
         return cameraman;
     }
     @DeleteMapping("api/cameraman/{id}")
-    String deleteCameraman(@PathVariable int id){
+    String deleteCameraman(@PathVariable Long id){
 
         return "Category delete";
 
@@ -62,7 +57,7 @@ public class ApiProviderController {
     }
 
     @GetMapping("api/agence/{id}")
-    AgenceDTO getAgence(@PathVariable int id){
+    AgenceDTO getAgence(@PathVariable Long id){
         return agenceService.get(id);
 
     }
@@ -72,17 +67,17 @@ public class ApiProviderController {
     }
 
     @PutMapping("api/agence/{id}")
-    String updateAgence(@RequestBody String body,@PathVariable int id){
+    String updateAgence(@RequestBody String body,@PathVariable Long id){
         String agence = body;
         return agence;
     }
     @DeleteMapping("api/agence/{id}")
-    String deleteAgence(@PathVariable int id){
+    String deleteAgence(@PathVariable Long id){
         return agenceService.delete(id);
     }
 
     @GetMapping("api/analyste/{id}")
-    AnalysteDTO getAnalyste(@PathVariable int id){
+    AnalysteDTO getAnalyste(@PathVariable Long id){
         return analysteService.get(id);
     }
 
@@ -92,12 +87,12 @@ public class ApiProviderController {
 
     }
     @PutMapping("api/analyste/{id}")
-    String updateAnalyste(@RequestBody String body,@PathVariable int id){
+    String updateAnalyste(@RequestBody String body,@PathVariable Long id){
         String analyste = body;
         return analyste;
     }
     @DeleteMapping("api/analyste/{id}")
-    String deleteAnalyste(@PathVariable int id){
+    String deleteAnalyste(@PathVariable Long id){
         return analysteService.delete(id);
 
     }
@@ -106,7 +101,7 @@ public class ApiProviderController {
         return analysteService.getAll();
     }
     @GetMapping("api/manager/{id}")
-    ManagerDTO getManager(@PathVariable int id){
+    ManagerDTO getManager(@PathVariable Long id){
         return managerService.get(id);
     }
     @PostMapping("api/manager")
@@ -114,12 +109,12 @@ public class ApiProviderController {
         return managerService.add(body);
     }
     @PutMapping("api/manager/{id}")
-    String updateManager(@RequestBody String body,@PathVariable int id){
+    String updateManager(@RequestBody String body,@PathVariable Long id){
         String manager = body;
         return manager;
     }
     @DeleteMapping("api/manager/{id}")
-    String deleteManager(@PathVariable int id){
+    String deleteManager(@PathVariable Long id){
         return managerService.delete(id);
     }
 
@@ -128,7 +123,7 @@ public class ApiProviderController {
         return managerService.getAll();
     }
     @GetMapping("api/marketing_advisor/{id}")
-    MarketingAdvisorDTO getMarketingAdvisor(@PathVariable int id){
+    MarketingAdvisorDTO getMarketingAdvisor(@PathVariable Long id){
         return marketingAdvisorService.get(id);
     }
     @PostMapping("api/marketing_advisor/{id}")
@@ -136,12 +131,12 @@ public class ApiProviderController {
         return marketingAdvisorService.add(body);
     }
     @PutMapping("api/marketing_advisor/{id}")
-    String updateMarketingAdvisor(@RequestBody String body,@PathVariable int id){
+    String updateMarketingAdvisor(@RequestBody String body,@PathVariable Long id){
         String marketingAdvisor = body;
         return marketingAdvisor;
     }
     @DeleteMapping("api/marketing_advisor/{id}")
-    String deleteMarketingAdvisor(@PathVariable int id){
+    String deleteMarketingAdvisor(@PathVariable Long id){
         return marketingAdvisorService.delete(id);
 
     }
@@ -152,23 +147,23 @@ public class ApiProviderController {
     }
 
     @GetMapping("api/wearable_society/{id}")
-    WearableSocietyDTO getWearableSociety(@PathVariable int id){
+    WearableSocietyDTO getWearableSociety(@PathVariable Long id){
         return wearableSocietyService.get(id);
     }
 
     @PostMapping("api/wearable_society/{id}")
-    WearableSocietyDTO addWearableSociety(@RequestBody WearableSocietyDTO body, @PathVariable int id){
+    WearableSocietyDTO addWearableSociety(@RequestBody WearableSocietyDTO body, @PathVariable Long id){
         WearableSocietyDTO wearableSociety = body;
         return wearableSociety;    }
 
     @PutMapping("api/wearable_society/{id}")
-    String updateWearableSociety(@RequestBody String body,@PathVariable int id){
+    String updateWearableSociety(@RequestBody String body,@PathVariable Long id){
         String wearableSociety = body;
         return wearableSociety;
 
     }
     @DeleteMapping("api/wearable_society/{id}")
-    String deleteWearableSociety(@PathVariable int id){
+    String deleteWearableSociety(@PathVariable Long id){
         return wearableSocietyService.delete(id);
     }
